fees details done

+++ resources/views/pages/masters/fees-heads.blade.php
@@ -1,6 +1,5 @@
 <x-main-card>
     <div class="">Fees Head</div>
-    <div class="w-full bg-gray-200" style="height: 1px;"></div>
     <div class="p-10 flex justify-around">
         <form class="w-full">
             <div class="relative z-0 w-1/4 mb-6 group">
@@ -11,18 +10,21 @@
                 <x-button-primary value="Save" />
             </div>
         </form>
-        
+
         <div class="flex flex-col w-full">
             <div class="overflow-x-auto sm:-mx-6 lg:-mx-8">
                 <div class="py-2 inline-block min-w-full sm:px-6 lg:px-8">
                     <div class="overflow-hidden">
                         <div class="w-full">
                             <button data-tooltip-target="tooltip-default" onclick="getFees()">
-                                <svg xmlns="http://www.w3.org/2000/svg" class="h-5 w-5 text-violet-500" fill="none" viewBox="0 0 24 24" stroke="currentColor" stroke-width="2">
-                                    <path stroke-linecap="round" stroke-linejoin="round" d="M4 4v5h.582m15.356 2A8.001 8.001 0 004.582 9m0 0H9m11 11v-5h-.581m0 0a8.003 8.003 0 01-15.357-2m15.357 2H15" />
-                                  </svg>
+                                <svg xmlns="http://www.w3.org/2000/svg" class="h-5 w-5 text-violet-500" fill="none"
+                                    viewBox="0 0 24 24" stroke="currentColor" stroke-width="2">
+                                    <path stroke-linecap="round" stroke-linejoin="round"
+                                        d="M4 4v5h.582m15.356 2A8.001 8.001 0 004.582 9m0 0H9m11 11v-5h-.581m0 0a8.003 8.003 0 01-15.357-2m15.357 2H15" />
+                                </svg>
                             </button>
-                            <div id="tooltip-default" role="tooltip" class="inline-block absolute invisible z-10 py-1 px-1.5 text-sm font-medium text-white bg-violet-400 rounded-lg shadow-sm opacity-0 transition-opacity duration-300 tooltip dark:bg-gray-700">
+                            <div id="tooltip-default" role="tooltip"
+                                class="inline-block absolute invisible z-10 py-1 px-1.5 text-sm font-medium text-white bg-violet-400 rounded-lg shadow-sm opacity-0 transition-opacity duration-300 tooltip dark:bg-gray-700">
                                 Refresh
                                 <div class="tooltip-arrow" data-popper-arrow></div>
                             </div>
@@ -38,7 +40,7 @@
                                     </th>
                                     <th scope="col" class="text-sm font-medium text-gray-900 px-6 py-4 text-left">
                                         Delete
-                                    </th>                                    
+                                    </th>
                                 </tr>
                             </thead>
                             <tbody>
@@ -112,27 +114,41 @@
     }
 
     function deleteFee(id) {
-        $.ajax({
-            type: "DELETE",
-            url: "{{route('master.deleteFee')}}",
-            data: {
-                id: id
-            },
-            dataType: "json",
-            beforeSend: function(){
-                $("#dele"+id).html("");
-                $("#dele"+id).append(`
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                $.ajax({
+                    type: "DELETE",
+                    url: "{{ route('master.deleteFee') }}",
+                    data: {
+                        id: id
+                    },
+                    dataType: "json",
+                    beforeSend: function() {
+                        $("#dele" + id).html("");
+                        $("#dele" + id).append(`
                     <svg role="status" class="w-4 h-4 mr-2 text-gray-200 animate-spin dark:text-gray-600 fill-red-600" viewBox="0 0 100 101" fill="none" xmlns="http://www.w3.org/2000/svg">
                         <path d="M100 50.5908C100 78.2051 77.6142 100.591 50 100.591C22.3858 100.591 0 78.2051 0 50.5908C0 22.9766 22.3858 0.59082 50 0.59082C77.6142 0.59082 100 22.9766 100 50.5908ZM9.08144 50.5908C9.08144 73.1895 27.4013 91.5094 50 91.5094C72.5987 91.5094 90.9186 73.1895 90.9186 50.5908C90.9186 27.9921 72.5987 9.67226 50 9.67226C27.4013 9.67226 9.08144 27.9921 9.08144 50.5908Z" fill="currentColor"/>
                         <path d="M93.9676 39.0409C96.393 38.4038 97.8624 35.9116 97.0079 33.5539C95.2932 28.8227 92.871 24.3692 89.8167 20.348C85.8452 15.1192 80.8826 10.7238 75.2124 7.41289C69.5422 4.10194 63.2754 1.94025 56.7698 1.05124C51.7666 0.367541 46.6976 0.446843 41.7345 1.27873C39.2613 1.69328 37.813 4.19778 38.4501 6.62326C39.0873 9.04874 41.5694 10.4717 44.0505 10.1071C47.8511 9.54855 51.7191 9.52689 55.5402 10.0491C60.8642 10.7766 65.9928 12.5457 70.6331 15.2552C75.2735 17.9648 79.3347 21.5619 82.5849 25.841C84.9175 28.9121 86.7997 32.2913 88.1811 35.8758C89.083 38.2158 91.5421 39.6781 93.9676 39.0409Z" fill="currentFill"/>
                     </svg>
                 `);
-            },
-            success: function (response) {
-                $("#tr"+id).remove();
+                    },
+                    success: function(response) {
+                        $("#tr" + id).remove();
+                    }
+
+                });
             }
-            
-        });
+        })
+
+
     }
 
     $("form").submit(function(e) {
@@ -172,7 +188,7 @@
                             </td>
                         </tr>`);
             },
-            error: function (xhr, status, error) {
+            error: function(xhr, status, error) {
                 let err = JSON.parse(xhr.responseText);
                 $(".loading").html("")
                 $(".loading").append(
